% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_color_palette.R
\name{get_color_palette}
\alias{get_color_palette}
\title{Generate a color palette for a vector of values}
\usage{
get_color_palette(
  x,
  theme,
  main_colors = cfg$themes[[theme]]$colors,
  cfg = loadConfig()
)
}
\arguments{
\item{x}{A vector of values for which a color palette is needed.}

\item{theme}{\link{character} Theme to apply to the plot.}

\item{main_colors}{A character vector of color names or hex codes to use as
the base palette.}

\item{cfg}{a config object. Configuration to use. See \link{loadConfig} for details}
}
\value{
A named character vector of colors, with names corresponding to the
unique values in \code{x}.
}
\description{
This function returns a named vector of colors corresponding to the unique
values in \code{x}. If the number of unique values is less than or equal to
the length of \code{main_colors}, it selects the first \code{n} colors from
\code{main_colors}. Otherwise, it generates a palette of \code{n} colors
interpolated from \code{main_colors}.
}
\examples{
get_color_palette(c("A", "B", "C"), main_colors = c("red", "green", "blue"))

}
