% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kplot-package.R
\docType{package}
\name{kplot-package}
\alias{kplot-package}
\title{kplot: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Kevin Bosirany Orlando \email{kevinbosirany@gmail.com} (\href{https://orcid.org/0009-0009-2784-3108}{ORCID})

}
\keyword{internal}
