% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kplot.R
\name{kplot}
\alias{kplot}
\title{Plot data}
\usage{
kplot(x, package = "ggplot", ...)
}
\arguments{
\item{x}{An object to be plotted.}

\item{package}{The plotting package to use (default is "ggplot").}

\item{...}{Additional arguments passed to the specific plotting function.}
}
\value{
A plot of the object \code{x}.
}
\description{
Entry point to plot data with different plotting back ends. \code{package}
selects the back end: \code{kplot(x, "ggplot", ...)} calls \code{\link[=kplot.ggplot]{kplot.ggplot()}}
(implemented by the \pkg{kggplot} package), \code{kplot(x, "tsstudio", ...)}
calls \code{\link[=kplot.tsstudio]{kplot.tsstudio()}}. Any function named \code{kplot.<package>} is a valid
back end.
}
