% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kplot-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{get_color_palette}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{kggplot}{\code{\link[kggplot]{get_color_palette}}}
}}
