% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_theme.R
\name{set_theme}
\alias{set_theme}
\alias{set_theme.ggplot}
\title{Set Plot Theme for Various Plot Objects}
\usage{
set_theme(x, ...)

\method{set_theme}{ggplot}(x, theme = NULL, scale, vars, ...)
}
\arguments{
\item{x}{A plot object (e.g., a ggplot object).}

\item{...}{Additional arguments passed to the plotting method.}

\item{theme}{\link{character} Theme to apply to the plot.}

\item{scale}{\link{character} A vector indicating which aesthetics to scale (e.g.,
"color", "fill").}

\item{vars}{\link{list} Variables to plot}
}
\value{
A plot object with the specified theme applied.
}
\description{
Generic function to set a plot theme for different types of plot objects.
}
