% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kplot.ggplot.R
\name{kplot.ggplot}
\alias{kplot.ggplot}
\alias{kplot.ggplot.data.frame}
\title{Plot data with ggplot}
\usage{
kplot.ggplot(x, ...)

\method{kplot.ggplot}{data.frame}(
  x,
  vars,
  type,
  labels = list(title = NULL, x = NULL, y = NULL, color = vars$color, fill = vars$fill),
  theme = NULL,
  ...
)
}
\arguments{
\item{x}{Data to plot.}

\item{...}{Additional arguments passed to the plotting method.}

\item{vars}{\link{list} Variables to plot}

\item{type}{\link{character} Type of plot to create (e.g., "point", "line",
"density", "cumFreq", "bar", "bar_dodge", "convexhull").}

\item{labels}{\link{list} Labels for the plot (title, x, y, color, fill).}

\item{theme}{\link{character} Theme to apply to the plot.}
}
\value{
A ggplot object.
}
\description{
Generic and method specific functions to plot data using ggplot2.
}
\examples{
data(iris)
kplot.ggplot(
  iris,
  vars = list(x = "Sepal.Length", y = "Sepal.Width", color = "Species"),
  type = "point",
  labels = list(
    title = "Iris Sepal Dimensions", x = "Sepal Length", y = "Sepal Width"
  )
)

}
