% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kplot.R
\name{kplot}
\alias{kplot}
\title{Plot data}
\usage{
kplot(x, package = "ggplot", ...)
}
\arguments{
\item{x}{An object to be plotted.}

\item{package}{The plotting package to use (default is "ggplot").}

\item{...}{Additional arguments passed to the specific plotting function.}
}
\value{
A plot of the object \code{x}.
}
\description{
This function is the main function to plot data using different plotting
packages.
}
